project2/requirements.txt
grpcio==1.58.0
grpcio-tools==1.58.0